+++ R/MetaAnalysis.R
@@ -39,8 +39,8 @@ doMetaAnalysis <- function(connectionDetails,
                            cacheFileName = NULL) {
 
   if (!is.null(diagnosticsFilter)) {
-    if (!(c("targetId", "comparatorId", "outcomeId", "analysisId",
-            "databaseId", "pass") %in% names(diagnosticsFilter))) {
+    if (!(all(c("targetId", "comparatorId", "outcomeId", "analysisId",
+            "databaseId", "pass") %in% names(diagnosticsFilter)))) {
       stop("Improperly formatted diagnostics filter")
     }
   }
@@ -140,17 +140,27 @@ doMaEffectType <- function(connectionDetails,
   allResults <- loadMainResults(connectionDetails, resultsDatabaseSchema,
                                  cacheFileName)
 
+  ncIds <- allResults %>% filter(trueEffectSize == 1) %>% pull(outcomeId) %>% unique()
+
   if (!is.null(diagnosticsFilter)) {
-    blind <- allResults %>%
-      left_join(diagnosticsFilter,
-                by = c("targetId",
-                       "comparatorId",
-                       "analysisId",
-                       "databaseId")) %>%
-      filter("pass") %>% select(-"pass")
+    diagnosticsFilter <- diagnosticsFilter %>%
+      filter(pass) %>%
+      select("targetId", "comparatorId", "analysisId", "databaseId", "outcomeId")
+
+    ncDiagnostics <- diagnosticsFilter %>%
+      distinct(targetId, comparatorId, analysisId, databaseId) %>%
+      merge(data.frame(outcomeId = ncIds))
+
+    blind <- rbind(diagnosticsFilter, ncDiagnostics)
+
+    allResults <- allResults %>%
+      inner_join(blind, by = c("targetId",
+                               "comparatorId",
+                               "analysisId",
+                               "outcomeId",
+                               "databaseId"))
   }
 
-  ncIds <- allResults %>% filter(trueEffectSize == 1) %>% pull(outcomeId) %>% unique()
   allResults$type[allResults$outcomeId %in% ncIds] <- "Negative control"
   allResults$type[is.na(allResults$type)] <- "Outcome of interest"
 
@@ -164,14 +174,12 @@ doMaEffectType <- function(connectionDetails,
   ParallelLogger::stopCluster(cluster)
   results <- do.call(rbind, results)
 
-  maName <- "Meta-analysis"
   results <- results %>% mutate(databaseId = maName) %>%
     select(-trueEffectSize,-type)
   colnames(results) <- SqlRender::camelCaseToSnakeCase(colnames(results))
 
   fileName <- file.path(maExportFolder, "cohort_method_result.csv")
   write.csv(results, fileName, row.names = FALSE, na = "")
-
 }
 
 computeGroupMetaAnalysis <- function(group,
@@ -269,7 +277,8 @@ computeSingleMetaAnalysis <- function(outcomeGroup) {
     maRow$mdrr <- exp(sqrt((zBeta + z1MinAlpha)^2/(totalEvents * pA * pB)))
   }
   maRow$databaseId <- "Meta-analysis"
-  maRow$sources <- paste(outcomeGroup$databaseId[order(outcomeGroup$databaseId)], collapse = ";")
+  maRow$sources <- paste(outcomeGroup$databaseId[order(outcomeGroup$databaseId)],
+                         collapse = ";")
   return(maRow)
 }